% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{acs_age_sex_race_ethnicity_vars}
\alias{acs_age_sex_race_ethnicity_vars}
\title{ACS variables for age, sex, race, and ethnicity}
\format{
A \code{\link[tibble]{tibble}} with 65 rows and 2 variables:
  \describe{ \item{variable}{ACS variable name}
  \item{description}{A description of who is present in the count}}
}
\usage{
acs_age_sex_race_ethnicity_vars
}
\description{
A two-column data set of the American Community Survey variable names and
their descriptions. Contains counts of various subdivisions of the population
based on age, sex, race, and ethnicity.
}
\details{
These variable names have been consistent throughout the existence of the ACS
from its beginning through 2020.

This data set is used to support \code{\link{synthetic_population}()}.
}
\seealso{
\code{\link{decennial_age_sex_race_ethnicity_vars}}
}
\keyword{datasets}
