% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centers.R
\name{areas_in_radius}
\alias{areas_in_radius}
\title{Make a tibble of census areas within a user-specified circle}
\usage{
areas_in_radius(
  geography,
  center = lon_lat_from_area(state = "DC"),
  radius = 5,
  units = "miles",
  measure_from = "center of population",
  year = 2010,
  distance_fun = geosphere::distVincentyEllipsoid
)
}
\arguments{
\item{geography}{The type of census areas that the resulting table will
contain. One of \code{c("state", "county", "tract", "block group")}.}

\item{center}{The longitude/latitude coordinates of the center of the circle.
A double vector of length 2 whose elements are finite numbers. Passed to
the \code{y} argument of \code{\link[geosphere:distm]{geosphere::distm()}}.

The first element is the longitude coordinate (positive for west, negative
for east). The second element is the latitude coordinate (positive for
north, negative for south).

The convenience function \code{\link[=lon_lat_from_area]{lon_lat_from_area()}} can be used to obtain the
longitude/latitude coordinates of the center of population of a
user-specified census area.

Defaults to the center of population of the District of Columbia according
to the 2010 decennial census.}

\item{radius}{A single, non-missing number specifying the radius of the
circle. Processed with \code{\link[=as.double]{as.double()}}. Defaults to 5.}

\item{units}{A single string specifying the units of \code{radius}. Passed to the
\code{value} argument of \code{\link[units:units]{units::set_units}}\code{(mode = "standard")}. Defaults to
\code{"miles"}.}

\item{measure_from}{Currently can only be \code{"center of population"}, the
default.}

\item{year}{Must be either 2000 or the default of 2010.}

\item{distance_fun}{Passed to the \code{fun} argument of \code{\link[geosphere:distm]{geosphere::distm()}}.
Defaults to \code{\link[geosphere:distVincentyEllipsoid]{geosphere::distVincentyEllipsoid}}, which results in the most
accurate measurement but is also the slowest.}
}
\value{
A \code{\link[tibble:tibble]{tibble}} with columns \code{geoid}, \code{longitude},
\code{latitude}, and \code{distance}. The \code{longitude} and \code{latitude} are the center
of population of the area, and \code{distance} is the number of \code{units} the
area's center of population is away from the coordinates given in \code{center}.
}
\description{
Returns a table containing all census areas whose centers of population are
within a radius of some center. To specify the center, the user can manually
enter longitude/latitude coordinates or use the helper function
\code{\link[=lon_lat_from_area]{lon_lat_from_area()}} to automatically grab the longitude/latitude
coordinates of the center of population of an area.
}
\details{
Centers of population are based on the 2010 decennial census. See
\url{https://www.census.gov/geographies/reference-files/time-series/geo/centers-population.2010.html}
Only states, counties, tracts, and block groups are currently supported.

Distances are determined with \code{\link[geosphere:distm]{geosphere::distm()}}.

Requires the packages \code{USpopcenters}, \code{geosphere}, and \code{units} to be
installed.
}
\examples{
# All states within 300 kilometers of the center of population of Manhattan
areas_in_radius(
  geography = "state",
  center = lon_lat_from_area(state = "NY", county = "New York"),
  radius = 300,
  units = "km"
)

# All census tracts within 24 miles of the four corners
areas_in_radius("tract", center = c(-109.0452, 36.9991), radius = 24)
}
\seealso{
\code{\link[=lon_lat_from_area]{lon_lat_from_area()}}
}
