% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_adi.R
\name{calculate_adi}
\alias{calculate_adi}
\title{Calculate ADI from census data.}
\usage{
calculate_adi(data, keep_indicators = FALSE)
}
\arguments{
\item{data}{Raw data obtained via
\code{tidycensus::\link[tidycensus]{get_acs}(output = "wide")} or
\code{tidycensus::\link[tidycensus]{get_decennial}(output = "wide")},
having the variables necessary to compute the indicators of the ADI, which
can be found in \code{sociome::\link{acs_vars}},
\code{sociome::decennial_vars_1990}, or
\code{sociome::decennial_vars_2000}.}

\item{keep_indicators}{Logical indicating whether or not to keep the
component indicators of the ADI as well as the original census variables
used to calculate them. Defaults to \code{FALSE}.}
}
\description{
Calculate the area deprivation index using decennial Census or American
Community Survey (ACS) variables.
}
\details{
This function discerns between ACS, year 2000 decennial census, and year 1990
decennial census data by checking for the existence of the population
variables of each (\code{B01003_001}, \code{P001001}, and \code{P0010001},
respectively).
}
