% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{dataset_year_geography_availability}
\alias{dataset_year_geography_availability}
\title{Working Combinations of \code{dataset}, \code{year}, and \code{geography} in \code{\link[=get_adi]{get_adi()}}}
\format{
A \code{\link[tibble:tibble]{tibble}} with 130 rows and 4 variables:
\itemize{
\item \strong{dataset}: Value that can be supplied to the \code{dataset} argument of \code{\link[=get_adi]{get_adi()}}.
\item \strong{year}: Value that can be supplied to the \code{year} argument of \code{\link[=get_adi]{get_adi()}}.
\item \strong{geography}: Value that can be supplied to the \code{geography} argument of \code{\link[=get_adi]{get_adi()}}. Note that \code{"zip code tabulation area"} is not represented here: just use \code{"zcta"}, please.
\item \strong{acs}: Logical, indicating that the combination of variables can yield ADIs and ADI-3s.
}
}
\usage{
dataset_year_geography_availability
}
\description{
A three-column data set of combinations of the \code{dataset}, \code{year}, and
\code{geography} arguments to \code{\link[=get_adi]{get_adi()}} that should successfully yield ADI and
ADI-3 estimates.
}
\seealso{
\code{\link[=get_adi]{get_adi()}}
}
\keyword{datasets}
