% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{decennial_age_sex_race_ethnicity_vars}
\alias{decennial_age_sex_race_ethnicity_vars}
\title{Decennial Census variables for age, sex, race, and ethnicity}
\format{
A \code{\link[tibble:tibble]{tibble}} with 195 rows and 3 variables:
\itemize{
\item \strong{year}: The year of the decennial census with which the
variable is associated.
\item \strong{variable}: ACS variable name.
\item \strong{description}: A description of who is present in the count.
}
}
\usage{
decennial_age_sex_race_ethnicity_vars
}
\description{
A three-column data set of the Decennial Census variable names, their
descriptions, and their decennial census year. Contains counts of various
subdivisions of the population based on age, sex, race, and ethnicity.
}
\details{
Currently, the 2000, 2010, and 2020 Decennial Census variables are available.

This data set is used to support \code{\link{synthetic_population}()}.
}
\seealso{
\code{\link[=synthetic_population]{synthetic_population()}} and
\code{\link{acs_age_sex_race_ethnicity_vars}}
}
\keyword{datasets}
