% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_adi.R
\name{mice.impute.pmm}
\alias{mice.impute.pmm}
\title{\code{mice.impute.pmm} from the \code{mice} package}
\description{
This is only visible in the sociome package out of necessity so that
imputation works.
}
\details{
See \code{\link[mice]{mice.impute.pmm}} for its actual documentation.

The reason that this function is in the visible exported namespace of sociome
is as follows:
\code{mice::mice()} internally calls one or more of the imputation method
functions in the \code{mice} package (in this case,
\code{mice.impute.pmm()}), but \code{mice::mice} looks for these functions in
an unconventional way: it looks only in the global environment and all its
parent environments. Therefore, the mice package must be attached in order
for \code{mice::mice()} to work. Instead of forcing the user to attach
\code{mice} whenever sociome needs to utilize \code{mice::mice()}, we elected
to import the imputation method function that we used
(\code{mice.impute.pmm()}) into the exported namespace of sociome. In effect,
this lets the user choose whether to attach the sociome package OR the
\code{mice} package (e.g., running \code{library(sociome)} or
\code{library(mice)}).
}
