% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_GEOIDs.R
\name{parse_GEOIDs}
\alias{parse_GEOIDs}
\title{parse_GEOIDs}
\usage{
parse_GEOIDs(GEOIDs)
}
\arguments{
\item{GEOIDs}{a character vector of GEOIDs. All elements of the vector must
contain the same number of digits so that they represent the same level of
geography.}
}
\value{
A list of class "GEOIDs", which contains the original vector of
  GEOIDs and a string indicating the level of geography that the
  GEOIDs represent (one of "state", "county", "tract", or "block group").
}
\description{
Scans a vector of GEOIDs for problems, and returns a list of class "GEOIDs"
to be used by get_reference_area()
}
