% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{state_geoids}
\alias{state_geoids}
\title{Vector of the two-digit GEOIDs for the 50 states, DC and Puerto Rico}
\format{
An object of class \code{character} of length 52.
}
\usage{
state_geoids
}
\description{
52-element named character vector of the two-digit GEOIDs of the 50 states,
DC and Puerto Rico. The names are the spelled-out names of each place.
}
\details{
These are derived from \link[tidycensus:fips_codes]{tidycensus::fips_codes}. Other territories not
mentioned here are simply filtered out.
}
\examples{
state_geoids
}
\keyword{datasets}
