% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{decennial_vars}
\alias{decennial_vars}
\title{Decennial census variable names for ADI calculation}
\format{A \code{\link[tibble]{tibble}} with 134 rows and 4 variables:
  \describe{ \item{variable}{Decennial census variable name}
  \item{sumfile}{The summary tape file of the decennial census variable}
  \item{year}{The year of the decennial census variable}
  \item{description}{Brief description of the data the variable contains}}}
\usage{
decennial_vars
}
\description{
A dataset of the decennial census variable names used to calculate the Area
Deprivation Index (ADI).
}
\seealso{
\code{\link{acs_vars}}
}
\keyword{datasets}
