% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_adi.R
\name{get_adi}
\alias{get_adi}
\title{get_adi}
\usage{
get_adi(geography = NULL, state = NULL, county = NULL,
  geoids = NULL, year = NULL, survey = "acs5", geometry = TRUE,
  key = NULL, ...)
}
\arguments{
\item{geography}{a character string denoting the type of location whose ADIs
you'd like to see (e.g., "region", "county", "block group"). Must be one of
the keywords accepted by the tidycensus::get_acs() function (see
\url{https://walkerke.github.io/tidycensus/articles/basic-usage.html#geography-in-tidycensus}
 for a full list of options).}

\item{state}{The state for which you are requesting ADI data. Not applicable
for certain values of `geography`, required for other values of
`geography`, and optional for still other values of `geography` (see
\url{https://walkerke.github.io/tidycensus/articles/basic-usage.html#geography-in-tidycensus}
 for more information). State names, postal codes, and FIPS codes are
accepted. Defaults to NULL.}

\item{county}{The county for which you are requesting ADI data. County names
and FIPS codes are accepted. Must be combined with a value supplied to
'state'. Not applicable for some values of `geography`, optional for other
values of `geography`, but never required (see
\code{\link[tidycensus]{census_api_key}} for more information).
https://walkerke.github.io/tidycensus/articles/basic-usage.html#geography-in-tidycensus
 for more information). Defaults to NULL.}

\item{geoids}{A character vector of GEOIDs. See details.}

\item{year}{The year, or endyear, of the ACS sample used to calculate the ADI
values. 2010 through 2016 are available. Defaults to 2016.}

\item{survey}{The ACS contains one-year, three-year, and five-year surveys
expressed as "acs1", "acs3", and "acs5". The default selection is "acs5."}

\item{key}{Your Census API key. Obtain one at
http://api.census.gov/data/key_signup.html. Not necessary if you have
already loaded your key with the tidycensus::census_api_key() function. See}
}
\value{
A tibble of three columns: GEOID of location, Name of location, ADI
  of location
}
\description{
Returns a tibble of user-specified locations in the United States along with
their area deprivation indices (ADIs), utilizing American Community Survey
data.
}
\details{
Elements of `geoids` can represent different levels of geography,
  but they all must be either 2 digits (for states), 5 digits (for counties),
  11 digits (for tracts) or 12 digits (for block groups). Be sure to put a
  leading zero where applicable.
}
\examples{
get_adi(geography = "tract", year = 2015, state = "OH", county = "Cuyahoga")
get_adi(geography = "region")

}
