% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geoids.R
\name{get_geoids}
\alias{get_geoids}
\title{Obtain GEOIDs of places}
\usage{
get_geoids(geography, state = NULL, county = NULL, geoid = NULL,
  year = 2010, geometry = FALSE, cache_tables = TRUE, key = NULL,
  ...)
}
\arguments{
\item{geography}{A character string denoting the level of census geography
whose GEOIDs you'd like to obtain. Must be one of \code{c("state",
"county", "tract", "block group", "block")}.}

\item{state}{A character string specifying states containing the desired
GEOIDs. Defaults to \code{NULL}. Can contain full state names, two-letter
state abbreviations, or two-digit FIPS codes/GEOIDs (must be vector of
strings, so use quotation marks and leading zeros if necessary). Must be
left as \code{NULL} blank if using the \code{geoid} parameter.}

\item{county}{A vector of character strings specifying the counties
containing the desired GEOIDs. Defaults to \code{NULL}. If specified, the
\code{state} parameter must contain exactly one state. County names and
three-digit FIPS codes are accepted (must contain strings, so use quotation
marks and leading zeros if necessary). Must be blank if using the
\code{geoid} parameter.}

\item{geoid}{A character vector of GEOIDs (use quotation marks and leading
zeros). Defaults to \code{NULL}. Must be blank if \code{state} and/or
\code{county} is used. Can contain different levels of geography (see
details).}

\item{year}{Single integer specifying the year of US Census data to use.
Defaults to 2010. Based on this year, data from the most recent decennial
census will be returned.}

\item{geometry, cache_tables, key}{See the descriptions of the arguments in
\code{\link{get_adi}()}.}

\item{...}{Additional arguments to be passed to
\code{tidycensus::\link[tidycensus]{get_decennial}()}. Not recommended; use
at your own risk.}
}
\description{
Returns a \code{tibble} of GEOIDs, names, and decennial census population of
user-specified locations.
}
\details{
This allows users to quickly obtain all GEOIDs in a specified location at a
specific level of geography without having to manually look them up somewhere
else.

This facilitates calls to \code{\link{get_adi}()} that involve somewhat
complicated reference areas.
}
\examples{
\donttest{
# Wrapped in \donttest{} because it requires a Census API key.

# Get all county GEOIDs for all Connecticut counties
get_geoids(geography = "county", state = "CT")
}
}
