% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_adi.R
\name{mice.impute.pmm}
\alias{mice.impute.pmm}
\title{Import mice.impute.pmm from mice package}
\usage{
mice.impute.pmm(y, ry, x, wy = NULL, donors = 5L, matchtype = 1L,
  ridge = 1e-05, ...)
}
\description{
This simply imports mice.impute.pmm from the mice package.
}
\details{
This is only here so that mice::mice() will work. It needs mice.impute.pmm(),
and it will not find it unless it is included as an exported function in
sociome.
}
