% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_adi.R
\name{mice.impute.pmm}
\alias{mice.impute.pmm}
\title{\code{mice.impute.pmm} from the \code{mice} package}
\description{
This is only visible in the sociome package out of necessity so that
imputation works.
}
\details{
See \code{\link[mice]{mice.impute.pmm}} for its actual documentation.
}
